v0.1 beta release No new code here, just some readme changes and a version bump.
From now on I'm going to push updates to a testing branch before they make it back to the main repository. Keep an eye out for that if you're interested.
Be warned that the testing branch might get code that's broken/incompatible sometimes. If you're worried about stability then stay on the master branch, which I'll only update after I've had a chance to test it as much as possible.