implements a function to restart the daemon by command or signal
Because sometimes tho communication between daemon and KB is corrupted
after resuming from Standby or suspend, a restart function is
implemented.
It first calls the quit() funtion, then it calls main() again with the
original parameter list.

There are two ways to restart the daemon:
- send the string "restart some-description-as-one-word" to the cmd-pipe
  (normally /dev/input/ckb1/cmd or /dev/input/ckb2/cmd, dependign on
  what device gets what usb number

- sending SIGUSR1 to the daemon-process (as root).

Later on, there may be a user interface in the client
for the first method.