add a cgeventtap to fix loss of modifier keys
This adds an event tap to the daemon's runloop so that whenever the
mouse moves we check the modifiers on the keyboard and make sure they
are added to the event rather than lost. This rectifies the loss of
modifier keys for things like cmd-click and alt-drag etc.

also add requested PR changes:

- Add comments to the code
- Cleanup code, move global variable declaration to top of file, cleanup
  spacing and indenting
- fixed issue where if an apple keyboard was also plugged in the
  modifier keys were being lost on mouse events