Fix init bug for indicator-keys
After a reboot of the GUI, the indicator keys have only responded
to settings in the Performance tab when one of the settings
has been changed.

The reason for this is in the initialization sequence to the daemon.
With the inotify command executed here,
the Notifychannel must obviously be preceded.

On the occasion a few methods were commented
according to doxygen standard.
Although they had nothing to do with the error,
they were analyzed during debugging.