Unlock coloring while playing macros with delays
As a user mentioned, the colorization of the keyboard was stopped
while a macro with delays is played.

This was because the synchronisation block in input.c was over the whole
play-loop for the macro.
Now we have an unlock / lock sequence around all usleep calling for delay.

This has only minimal impact on the timeing behavior.
Test was a long running macro inclusive startings
"time cat" at the beginning and sending CTRL-D at the end.
Colorization was "shimmer" all over the keyboard,
so heavy load between GUI and daemon.

With this long macro before this commit, time-cmd gives:
real    0m40.087s   (between 40.080 and 40.087 in different runs)
user    0m0.000s
sys     0m0.000s

after changing the blocking for this critical section we get:
real    0m40.116s
user    0m0.000s
sys     0m0.000s