Set DPI flag in a way to minimize number of commands.
We use an intermediate enabled flag only when necessary - that is, when enabling the new stage but disabling the current one, we need an intermediate set of flags in which both stages are enabled.

This commit also contains a small change to cache the DPI settings when switching DPI stages. Also added some comments.