Update udev rules to create symlinks for devices
this change creates symlinks in /dev/input/by-id to the event objects for each device. Each ckb device has one mouse and one keyboard event that have the same name and little to distinguish them from each other, so to handle this I appended the EV count to the end of the name. 7 is mouse and 120003 is keyboard. Since these event objects are technically not the actually device but more of a fake device to facilitate remapping I felt it was appropriate to leave `ckb` at the front of the name rather than the traditional `usb`.