Improve safety of signal handling
This commit introduces a safer signal handling routine by splitting this
workflow into catching and handling of signals. In the catching routine
no unsafe function call is allowed (signal-safety(7)).

The general idea is generate a local socketpair (acting as an anonymous
pipe) that is written to with the caught signal type in the direct
sighandler routine. The calling of the routine that actually handles the
interrupts is elevated to the main loops, where the pipe is monitored.
As soon as data is available on the pipe it is read and the
sighandling-routine is called manually. This allows for greater freedom
of function calls to handle the signal with.

- Create anonymous socket pair for sending/receiving intent, only setup
  signal handlers if the socket pair exists.
- Call `write` to sender side of socket pair in signal-catching
  function.
- Replace `printf` with signal-safe `write` function in nested
  `ignore_signal` handler.
- [usb] Declare the socket pair and the exit condition as externally
  defined for calling them from inside the main loops.
- [macOS] Setup and read from receiver side via a core foundation socket
  object, that is registered as a CF Run Loop Source, deconstructed into
  a native socket and read from in the appropriate handler.
- [linux] Check whether data is available on the pipe when `select`
  returns -1, indicating that there has been an error (e.g. a signal).
  Manually call the signal handling routine with the read data if there
  is.