Allow for faster macro calls
The current code limits one macro invocation per 500 ms, even if the
button the macro is bound to is release and pressed again.  This breaks
convention---typically, users expect that releasing and depressing a
button will cause the effect to be applied again.  The repeat delay is
only expected if the button is held down continuously.

Centralizes all macro key-press checks to inputupdate_keys, which now
signals the state of the keys in the triggered field of the keymacro
struct.  Access to this field is protected by the imutex, and changes
are signaled to waiting threads via macroint pthread_cond_t, accessible
via the mintvar() macro.

Key up and key down events are counted, so keypresses are not lost.

Signed-off-by: Antonio Russo <aerusso@aerusso.net>