Protect play_macro() threads from macro changes
play_macro() currently accesses the keymacro data structure passed to it
at creation time.  However, this structure may be changed if
cmd_macro() is called.

Pass the macro actions and actioncount directly in parameter_t, curing
the race condition if a macro is triggered, and the macro is redefined.

Furthermore, signal if macro has been redefined by informing the running
thread that it has been aborted.  Delegate the freeing of macro.actions
to the running thread.

Signed-off-by: Antonio Russo <aerusso@aerusso.net>