Split bragi keyboard input
This is needed to get SW key input in the K57 WL and possibly others
as unfortunately the header "01 02" in bragi means device 1
key input, but it just means "NKRO input" in NXP and legacy.

Perhaps the size check could be moved to the top, but there's no
reason to risk breaking older devices.